backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.2",
    "openpyxl>=3.1.0",
    "pandas>=2.0.0",
    "python-dotenv>=1.2.2",
    "requests>=2.31.0",
    "supabase>=2.28.3",
    "uvicorn[standard]>=0.42.0",
]
